everything working, just need to write documentation

+++ src/main/java/lb_backend/Note.java
@@ -1,12 +1,9 @@
 package lb_backend;
 
-/**
- * Created by Nate on 1/19/2017.
- */
 public class Note {
-    public Integer id;
+    public Long id;
     public String body;
-      public Note(Integer id, String body) {
+      public Note(Long id, String body) {
         this.id = id;
         this.body = body;
     }
